+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/configuration/JsonLdConfiguration.java
@@ -51,8 +51,9 @@ public class JsonLdConfiguration {
     public static final String NAMESPACE_EDC_PARTICIPANT_ID = NAMESPACE_EDC + "participantId";
     public static final String NAMESPACE_EDC_ID = NAMESPACE_EDC + "id";
     public static final String NAMESPACE_TRACTUSX = "https://w3id.org/tractusx/v0.0.1/ns/";
-    public static final String NAMESPACE_DCT = "https://purl.org/dc/terms/";
+    public static final String NAMESPACE_DCT = "http://purl.org/dc/terms/";
     public static final String JSON_LD_OBJECT_MAPPER = "jsonLdObjectMapper";
+    public static final String NAMESPACE_CATENAX_POLICY = "https://w3id.org/catenax/policy/";
 
     @Bean /* package */ TitaniumJsonLd titaniumJsonLd(final Monitor monitor) {
         final TitaniumJsonLd titaniumJsonLd = new TitaniumJsonLd(monitor);
@@ -62,6 +63,7 @@ public class JsonLdConfiguration {
         titaniumJsonLd.registerNamespace("edc", NAMESPACE_EDC);
         titaniumJsonLd.registerNamespace("dcat", JsonLdConfiguration.NAMESPACE_DCAT);
         titaniumJsonLd.registerNamespace("dspace", NAMESPACE_DSPACE);
+        titaniumJsonLd.registerNamespace("cx-policy", NAMESPACE_CATENAX_POLICY);
         return titaniumJsonLd;
     }