+++ irs-edc-client/src/test/java/org/eclipse/tractusx/irs/edc/client/transformer/JsonObjectToIrsPolicyTransformerTest.java
@@ -59,7 +59,7 @@ class JsonObjectToIrsPolicyTransformerTest {
                 "policy": {
                     "odrl:permission": [
                         {
-                            "odrl:action": "USE",
+                            "odrl:action": "use",
                             "odrl:constraint": {
                                 "odrl:and": [
                                     {