+++ irs-policy-store/src/main/java/org/eclipse/tractusx/irs/policystore/models/CreatePolicyRequest.java
@@ -76,7 +76,7 @@ public record CreatePolicyRequest(
                     "@type": "Policy",
                     "odrl:permission": [
                         {
-                            "odrl:action": "USE",
+                            "odrl:action": "use",
                             "odrl:constraint": {
                                 "odrl:and": [
                                     {